restructure code for more readability
