+++ src/StatusAggregator/Collector/ManualStatusChangeCollectorProcessor.cs
@@ -4,12 +4,13 @@
 using Microsoft.Extensions.Logging;
 using NuGet.Jobs.Extensions;
 using NuGet.Services.Status.Table.Manual;
+using StatusAggregator.Manual;
 using StatusAggregator.Table;
 using System;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace StatusAggregator.Manual
+namespace StatusAggregator.Collector
 {
     public class ManualStatusChangeCollectorProcessor : IEntityCollectorProcessor
     {
