+++ src/StatusAggregator/StatusAggregator.csproj
@@ -46,24 +46,24 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="EntityFactoryAggregator.cs" />
-    <Compile Include="ComponentAffectingEntityUpdater.cs" />
-    <Compile Include="EntityAggregationUpdateHandler.cs" />
-    <Compile Include="EntityCollector.cs" />
-    <Compile Include="EventFactory.cs" />
-    <Compile Include="IAggregatedEntityFactory.cs" />
-    <Compile Include="IComponentAffectingEntityUpdateHandler.cs" />
-    <Compile Include="IComponentAffectingEntityUpdateListener.cs" />
-    <Compile Include="IEntityAggregationLinkListener.cs" />
-    <Compile Include="IEntityCollector.cs" />
-    <Compile Include="IEntityCollectorProcessor.cs" />
-    <Compile Include="IEntityFactory.cs" />
-    <Compile Include="IComponentAffectingEntityUpdater.cs" />
-    <Compile Include="IncidentEntityCollectorProcessor.cs" />
-    <Compile Include="IncidentEntityUpdateHandler.cs" />
-    <Compile Include="IncidentFactory.cs" />
-    <Compile Include="IncidentGroupFactory.cs" />
-    <Compile Include="IncidentGroupLinkListener.cs" />
+    <Compile Include="Factory\EntityFactoryAggregator.cs" />
+    <Compile Include="Update\ComponentAffectingEntityUpdater.cs" />
+    <Compile Include="Update\EntityAggregationUpdateHandler.cs" />
+    <Compile Include="Collector\EntityCollector.cs" />
+    <Compile Include="Factory\EventFactory.cs" />
+    <Compile Include="Factory\IAggregatedEntityFactory.cs" />
+    <Compile Include="Update\IComponentAffectingEntityUpdateHandler.cs" />
+    <Compile Include="Update\IComponentAffectingEntityUpdateListener.cs" />
+    <Compile Include="Update\IEntityAggregationLinkListener.cs" />
+    <Compile Include="Collector\IEntityCollector.cs" />
+    <Compile Include="Collector\IEntityCollectorProcessor.cs" />
+    <Compile Include="Factory\IEntityFactory.cs" />
+    <Compile Include="Update\IComponentAffectingEntityUpdater.cs" />
+    <Compile Include="Collector\IncidentEntityCollectorProcessor.cs" />
+    <Compile Include="Update\IncidentEntityUpdateHandler.cs" />
+    <Compile Include="Factory\IncidentFactory.cs" />
+    <Compile Include="Factory\IncidentGroupFactory.cs" />
+    <Compile Include="Update\IncidentGroupLinkListener.cs" />
     <Compile Include="Manual\AddStatusEventManualChangeHandler.cs" />
     <Compile Include="Manual\AddStatusMessageManualChangeHandler.cs" />
     <Compile Include="Manual\DeleteStatusEventManualChangeHandler.cs" />
@@ -72,7 +72,7 @@
     <Compile Include="Manual\EditStatusMessageManualChangeHandler.cs" />
     <Compile Include="Manual\IManualStatusChangeHandler.cs" />
     <Compile Include="Manual\ManualStatusChangeHandler.cs" />
-    <Compile Include="Manual\ManualStatusChangeCollectorProcessor.cs" />
+    <Compile Include="Collector\ManualStatusChangeCollectorProcessor.cs" />
     <Compile Include="Manual\ManualStatusChangeUtility.cs" />
     <Compile Include="IMessageBuilder.cs" />
     <Compile Include="MessageBuilder.cs" />
@@ -94,7 +94,7 @@
     <Compile Include="Parse\AggregateIncidentParser.cs" />
     <Compile Include="Parse\IAggregateIncidentParser.cs" />
     <Compile Include="IStatusExporter.cs" />
-    <Compile Include="MessageEventUpdateListener.cs" />
+    <Compile Include="Update\MessageEventUpdateListener.cs" />
     <Compile Include="StatusExporter.cs" />
     <Compile Include="Parse\EnvironmentPrefixIncidentParser.cs" />
     <Compile Include="Parse\IIncidentParser.cs" />
