+++ src/StatusAggregator/Update/EntityAggregationUpdateHandler.cs
@@ -9,7 +9,7 @@ using NuGet.Jobs.Extensions;
 using NuGet.Services.Status.Table;
 using StatusAggregator.Table;
 
-namespace StatusAggregator
+namespace StatusAggregator.Update
 {
     public class EntityAggregationUpdateHandler<TEntityAggregation, TAggregatedEntity> 
         : IComponentAffectingEntityUpdateHandler<TEntityAggregation>
