+++ src/StatusAggregator/Update/IComponentAffectingEntityUpdateHandler.cs
@@ -5,16 +5,11 @@ using System;
 using System.Threading.Tasks;
 using NuGet.Services.Status.Table;
 
-namespace StatusAggregator
+namespace StatusAggregator.Update
 {
-    public interface IComponentAffectingEntityUpdater
-    {
-        Task UpdateAllActive(DateTime cursor);
-    }
-
-    public interface IComponentAffectingEntityUpdater<T>
+    public interface IComponentAffectingEntityUpdateHandler<T>
         where T : ComponentAffectingEntity
     {
-        Task<bool> Update(T entity, DateTime cursor);
+        Task<bool> Update(T groupEntity, DateTime cursor);
     }
 }
